refactor: 테스크 모달창 리팩토링 - 컴포넌트 분리
