+++ frontend/src/components/List/PreTaskList/style.ts
@@ -1,0 +1,21 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+    display: flex;
+    flex-direction: column;
+    width: 375px;
+    margin-left: 130px;
+`;
+
+export const PreTask = styled.div`
+    display: flex;
+    width: 375px;
+    align-items: center;
+`;
+
+export const DeleteButton = styled.img`
+    width: 15px;
+    height: 15px;
+    margin-right: 12px;
+    cursor: pointer;
+`;
